Move LocaleSwitcher to components/. Add tests for LocaleSwitcher and staticPageLoader
